common_script.sh: delete if $1=y & nuke colors.sh script

+++ tools/common_script.sh
@@ -10,7 +10,10 @@
 PROJECT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." >/dev/null && pwd )"
 
 # Text format
-source $PROJECT_DIR/tools/colors.sh
+bold=$(tput bold)
+cyan='\033[0;36m'
+nocol='\033[0m'
+red='\033[0;31m'
 
 # Create $PROJECT_DIR/working directory if it does not exist
 if [ ! -d $PROJECT_DIR/working ]; then
@@ -18,4 +21,6 @@ if [ ! -d $PROJECT_DIR/working ]; then
 fi
 
 # clean up
-rm -rf $PROJECT_DIR/working/*
+if [ "$1" = "y" ]; then
+	rm -rf $PROJECT_DIR/working/*
+fi
